Upgrader: ignore module compatibility.
Distribution JSONs come with a field 'compatibility', which wants
to give an idea about which migration module version is needed to
do a particular migration.

That's pointless, because even if incompatibilities come up, also
a new migration module version will be released, so a mismatch
is pretty unlikely. Also, there are currently no combinations not
working.